// frontend/src/shared/api/client/testConstructorAPI.schemas.ts
export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  single_choice: 'single_choice',
  multiple_choice: 'multiple_choice',
  text: 'text',
  essay: 'essay',
} as const;